Add dashboard properties panel and recent payments table
New GET /api/v1/payments/recent (manager-scoped, newest first) backs the
payments table. Sidebar group renamed to Main menu with Settings pinned to
the bottom.

Also fixes two pre-existing dashboard-chart defects surfaced by seeding real
data: monthly_income.amount serialised as a Decimal string while the frontend
type declared a number, and the Recharts entry animation left every bar at
zero height under React StrictMode.

+++ backend/app/schemas/stats.py
@@ -5,7 +5,10 @@ from pydantic import BaseModel
 
 class MonthlyIncome(BaseModel):
     month: str
-    amount: Decimal
+    # A float, not Decimal: Pydantic renders Decimal as a JSON string and the
+    # chart cannot plot strings. The money totals above stay Decimal - they are
+    # only ever displayed as text.
+    amount: float
 
 
 class DashboardStats(BaseModel):
